read/write ready looking good.
