+++ src/Makefile.in
@@ -67,6 +67,7 @@ SRCS = \
 	common/io/bufferRead.c \
 	common/io/bufferWrite.c \
 	common/io/client.c \
+	common/io/fd.c \
 	common/io/fdRead.c \
 	common/io/fdWrite.c \
 	common/io/filter/buffer.c \
