+++ src/common/io/socket/common.h
@@ -21,9 +21,4 @@ void sckOptionSet(int fd);
 // Connect socket to an IP address
 void sckConnect(int fd, const String *host, unsigned int port, const struct addrinfo *hostAddress, TimeMSec timeout);
 
-// Wait until the socket is ready to read/write or timeout
-bool sckReady(int fd, bool read, bool write, TimeMSec timeout);
-bool sckReadyRead(int fd, TimeMSec timeout);
-bool sckReadyWrite(int fd, TimeMSec timeout);
-
 #endif
